requirements.txt
flask
pymongo
beautifulsoup4
feedparser
requests
argon2-cffi
flask-limiter
